:sparkles: move admin advertise controllers inside module
