+++ modules/advertise/src/Http/Controllers/Admin/CategoryValueController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Controllers\Admin\Advertise;
+namespace Modules\Advertise\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\StoreCategoryValueRequest;
